requirements.txt
discord.py==2.3.2
praw==7.7.1
flask==3.0.3
pandas==2.2.2
PIL==9.0.1